+++ inc/core/templates/breadcrumbs.php
@@ -26,7 +26,12 @@ if (!function_exists('extrachill_breadcrumbs')) {
         // Allow plugins to override the default breadcrumb trail
         $custom_trail = apply_filters('extrachill_breadcrumbs_override_trail', '');
         if (!empty($custom_trail)) {
-            echo $custom_trail;
+            // Auto-wrap simple span trails with network dropdown
+            if ( preg_match( '/^<span>(.+)<\/span>$/', $custom_trail, $matches ) ) {
+                echo extrachill_network_dropdown( $matches[1] );
+            } else {
+                echo $custom_trail;
+            }
         } else {
             // Original breadcrumb logic
             if (is_single() && is_singular('post')) {